Add some doc around inset aware layouts
Fixes #366
