+++ sample/src/main/java/com/google/accompanist/sample/insets/InsetAwareLayouts.kt
@@ -1,0 +1,94 @@
+/*
+ * Copyright 2021 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.accompanist.sample.insets
+
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.material.BottomNavigation
+import androidx.compose.material.BottomNavigationDefaults
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.material.TopAppBar
+import androidx.compose.material.contentColorFor
+import androidx.compose.material.primarySurface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import com.google.accompanist.insets.navigationBarsPadding
+import com.google.accompanist.insets.statusBarsPadding
+
+/**
+ * A wrapper around [TopAppBar] which uses [Modifier.statusBarsPadding] to shift the app bar's
+ * contents down, but still draws the background behind the status bar too.
+ */
+@Composable
+fun InsetAwareTopAppBar(
+    title: @Composable () -> Unit,
+    modifier: Modifier = Modifier,
+    navigationIcon: @Composable (() -> Unit)? = null,
+    actions: @Composable RowScope.() -> Unit = {},
+    backgroundColor: Color = MaterialTheme.colors.primarySurface,
+    contentColor: Color = contentColorFor(backgroundColor),
+    elevation: Dp = 4.dp
+) {
+    Surface(
+        color = backgroundColor,
+        elevation = elevation,
+        modifier = modifier
+    ) {
+        TopAppBar(
+            title = title,
+            navigationIcon = navigationIcon,
+            actions = actions,
+            backgroundColor = Color.Transparent,
+            contentColor = contentColor,
+            elevation = 0.dp,
+            modifier = Modifier
+                .statusBarsPadding()
+                .navigationBarsPadding(bottom = false)
+        )
+    }
+}
+
+/**
+ * A wrapper around [BottomNavigation] which uses [Modifier.navigationBarsPadding] to shift
+ * the contents away from the system navigation bar, but still draws the background
+ * behind the system bar.
+ */
+@Composable
+fun InsetAwareBottomNavigation(
+    modifier: Modifier = Modifier,
+    backgroundColor: Color = MaterialTheme.colors.primarySurface,
+    contentColor: Color = contentColorFor(backgroundColor),
+    elevation: Dp = BottomNavigationDefaults.Elevation,
+    content: @Composable RowScope.() -> Unit
+) {
+    Surface(
+        color = backgroundColor,
+        elevation = elevation,
+        modifier = modifier
+    ) {
+        BottomNavigation(
+            backgroundColor = Color.Transparent,
+            contentColor = contentColor,
+            elevation = 0.dp,
+            modifier = Modifier.navigationBarsPadding(),
+            content = content
+        )
+    }
+}
